+++ src/unix/notbsd/linux/musl/b32/mips.rs
@@ -302,4 +302,4 @@ pub const TIOCMSET: ::c_ulong = 0x741D;
 pub const FIONREAD: ::c_ulong = 0x467F;
 pub const TIOCCONS: ::c_ulong = 0x80047478;
 
-pub const SYS_gettid: ::c_int = 4222;   // Valid for O32
+pub const SYS_gettid: ::c_long = 4222;   // Valid for O32
