+++ src/unix/notbsd/linux/other/b32/arm.rs
@@ -19,4 +19,4 @@ pub const SO_SNDTIMEO: ::c_int = 21;
 pub const FIOCLEX: ::c_ulong = 0x5451;
 pub const FIONBIO: ::c_ulong = 0x5421;
 
-pub const SYS_gettid: ::c_int = 224;
+pub const SYS_gettid: ::c_long = 224;
